Split `main` and `wasm` logic.
