+++ examples/download/src/lib.rs
@@ -1,7 +1,5 @@
 use std::path::PathBuf;
 
-#[cfg(not(target_arch = "wasm32"))]
-use clap::Parser;
 use peace::{
     resources::{
         resources_type_state::{
@@ -42,216 +40,12 @@ mod download_state_diff_fn_spec;
 mod file_state;
 mod file_state_diff;
 
-#[cfg(not(target_arch = "wasm32"))]
-pub fn run() -> Result<(), DownloadError> {
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .thread_name("main")
-        .thread_stack_size(3 * 1024 * 1024)
-        .enable_io()
-        .enable_time()
-        .build()
-        .map_err(DownloadError::TokioRuntimeInit)?;
-
-    let DownloadArgs { command } = DownloadArgs::parse();
-    runtime.block_on(async {
-        match command {
-            DownloadCommand::Status { url, dest } => {
-                let graph = setup_graph(url, dest).await?;
-                let resources = graph.setup(Resources::new()).await?;
-                status(&graph, resources).await?;
-            }
-            DownloadCommand::Desired { url, dest } => {
-                let graph = setup_graph(url, dest).await?;
-                let resources = graph.setup(Resources::new()).await?;
-                desired(&graph, resources).await?;
-            }
-            DownloadCommand::Diff { url, dest } => {
-                let graph = setup_graph(url, dest).await?;
-                let resources = graph.setup(Resources::new()).await?;
-                diff(&graph, resources).await?;
-            }
-            DownloadCommand::EnsureDry { url, dest } => {
-                let graph = setup_graph(url, dest).await?;
-                let resources = graph.setup(Resources::new()).await?;
-                ensure_dry(&graph, resources).await?;
-            }
-            DownloadCommand::Ensure { url, dest } => {
-                let graph = setup_graph(url, dest).await?;
-                let resources = graph.setup(Resources::new()).await?;
-                ensure(&graph, resources).await?;
-            }
-        }
-
-        Ok::<_, DownloadError>(())
-    })
-}
-
 #[cfg(target_arch = "wasm32")]
-use wasm_bindgen::prelude::*;
-
+mod wasm;
 #[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-extern "C" {
-    #[wasm_bindgen(js_namespace = console)]
-    fn log(s: &str);
-}
+use wasm::stdout_write;
 
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen(getter_with_clone)]
-pub struct GraphAndContent(
-    peace::rt_model::FullSpecGraph<DownloadError>,
-    std::collections::HashMap<PathBuf, String>,
-);
-
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-impl GraphAndContent {
-    /// Returns the content of the hashmap.
-    #[wasm_bindgen]
-    pub fn contents(&self) -> Result<JsValue, JsValue> {
-        JsValue::from_serde(&self.1).map_err(|e| JsValue::from_str(&format!("{e}")))
-    }
-}
-
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-pub async fn wasm_setup(url: String, name: String) -> Result<GraphAndContent, JsValue> {
-    std::panic::set_hook(Box::new(console_error_panic_hook::hook));
-
-    setup_graph(
-        Url::parse(&url).expect("Failed to parse URL."),
-        std::path::PathBuf::from(name),
-    )
-    .await
-    .map(|graph| async move {
-        let mut resources = graph
-            .setup(Resources::new())
-            .await
-            .map_err(|e| JsValue::from_str(&format!("{e}")))?;
-        let content = resources
-            .remove::<std::collections::HashMap<PathBuf, String>>()
-            .ok_or(JsValue::from_str(
-                "Resources did not contain content HashMap.",
-            ))?;
-        Ok(GraphAndContent(graph, content))
-    })
-    .map_err(|e| JsValue::from_str(&format!("{e}")))?
-    .await
-}
-
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-pub async fn wasm_status(graph_and_content: GraphAndContent) -> Result<GraphAndContent, JsValue> {
-    let GraphAndContent(graph, content) = graph_and_content;
-    let mut resources = graph
-        .setup(Resources::new())
-        .await
-        .map_err(|e| JsValue::from_str(&format!("{e}")))?;
-    resources.insert(content);
-
-    let mut resources = status(&graph, resources.into())
-        .await
-        .map_err(|e| JsValue::from_str(&format!("{e}")))?;
-
-    let content = resources
-        .remove::<std::collections::HashMap<PathBuf, String>>()
-        .ok_or(JsValue::from_str(
-            "Resources did not contain content HashMap.",
-        ))?;
-    Ok(GraphAndContent(graph, content))
-}
-
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-pub async fn wasm_desired(graph_and_content: GraphAndContent) -> Result<GraphAndContent, JsValue> {
-    let GraphAndContent(graph, content) = graph_and_content;
-    let mut resources = graph
-        .setup(Resources::new())
-        .await
-        .map_err(|e| JsValue::from_str(&format!("{e}")))?;
-    resources.insert(content);
-
-    let mut resources = desired(&graph, resources.into())
-        .await
-        .map_err(|e| JsValue::from_str(&format!("{e}")))?;
-
-    let content = resources
-        .remove::<std::collections::HashMap<PathBuf, String>>()
-        .ok_or(JsValue::from_str(
-            "Resources did not contain content HashMap.",
-        ))?;
-    Ok(GraphAndContent(graph, content))
-}
-
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-pub async fn wasm_diff(graph_and_content: GraphAndContent) -> Result<GraphAndContent, JsValue> {
-    let GraphAndContent(graph, content) = graph_and_content;
-    let mut resources = graph
-        .setup(Resources::new())
-        .await
-        .map_err(|e| JsValue::from_str(&format!("{e}")))?;
-    resources.insert(content);
-
-    let mut resources = diff(&graph, resources.into())
-        .await
-        .map_err(|e| JsValue::from_str(&format!("{e}")))?;
-
-    let content = resources
-        .remove::<std::collections::HashMap<PathBuf, String>>()
-        .ok_or(JsValue::from_str(
-            "Resources did not contain content HashMap.",
-        ))?;
-    Ok(GraphAndContent(graph, content))
-}
-
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-pub async fn wasm_ensure_dry(
-    graph_and_content: GraphAndContent,
-) -> Result<GraphAndContent, JsValue> {
-    let GraphAndContent(graph, content) = graph_and_content;
-    let mut resources = graph
-        .setup(Resources::new())
-        .await
-        .map_err(|e| JsValue::from_str(&format!("{e}")))?;
-    resources.insert(content);
-
-    let mut resources = ensure_dry(&graph, resources.into())
-        .await
-        .map_err(|e| JsValue::from_str(&format!("{e}")))?;
-
-    let content = resources
-        .remove::<std::collections::HashMap<PathBuf, String>>()
-        .ok_or(JsValue::from_str(
-            "Resources did not contain content HashMap.",
-        ))?;
-    Ok(GraphAndContent(graph, content))
-}
-
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-pub async fn wasm_ensure(graph_and_content: GraphAndContent) -> Result<GraphAndContent, JsValue> {
-    let GraphAndContent(graph, content) = graph_and_content;
-    let mut resources = graph
-        .setup(Resources::new())
-        .await
-        .map_err(|e| JsValue::from_str(&format!("{e}")))?;
-    resources.insert(content);
-
-    let mut resources = ensure(&graph, resources.into())
-        .await
-        .map_err(|e| JsValue::from_str(&format!("{e}")))?;
-
-    let content = resources
-        .remove::<std::collections::HashMap<PathBuf, String>>()
-        .ok_or(JsValue::from_str(
-            "Resources did not contain content HashMap.",
-        ))?;
-    Ok(GraphAndContent(graph, content))
-}
-
-async fn setup_graph(
+pub async fn setup_graph(
     url: Url,
     dest: PathBuf,
 ) -> Result<FullSpecGraph<DownloadError>, DownloadError> {
@@ -261,7 +55,7 @@ async fn setup_graph(
     Ok(graph)
 }
 
-async fn status(
+pub async fn status(
     graph: &FullSpecGraph<DownloadError>,
     resources: Resources<SetUp>,
 ) -> Result<Resources<WithStates>, DownloadError> {
@@ -275,7 +69,7 @@ async fn status(
     Ok(resources)
 }
 
-async fn desired(
+pub async fn desired(
     graph: &FullSpecGraph<DownloadError>,
     resources: Resources<SetUp>,
 ) -> Result<Resources<WithStatesDesired>, DownloadError> {
@@ -289,7 +83,7 @@ async fn desired(
     Ok(resources)
 }
 
-async fn diff(
+pub async fn diff(
     graph: &FullSpecGraph<DownloadError>,
     resources: Resources<SetUp>,
 ) -> Result<Resources<WithStateDiffs>, DownloadError> {
@@ -303,8 +97,7 @@ async fn diff(
     Ok(resources)
 }
 
-#[cfg_attr(target_arch = "wasm32", allow(dead_code))]
-async fn ensure_dry(
+pub async fn ensure_dry(
     graph: &FullSpecGraph<DownloadError>,
     resources: Resources<SetUp>,
 ) -> Result<Resources<EnsuredDry>, DownloadError> {
@@ -318,8 +111,7 @@ async fn ensure_dry(
     Ok(resources)
 }
 
-#[cfg_attr(target_arch = "wasm32", allow(dead_code))]
-async fn ensure(
+pub async fn ensure(
     graph: &FullSpecGraph<DownloadError>,
     resources: Resources<SetUp>,
 ) -> Result<Resources<Ensured>, DownloadError> {
@@ -334,18 +126,12 @@ async fn ensure(
 }
 
 #[cfg(not(target_arch = "wasm32"))]
-async fn stdout_write(s: &str) -> Result<(), DownloadError> {
+pub async fn stdout_write(s: &str) -> Result<(), DownloadError> {
     let mut stdout = io::stdout();
     stdout
         .write_all(s.as_bytes())
         .await
         .map_err(DownloadError::StdoutWrite)
-}
-
-#[cfg(target_arch = "wasm32")]
-async fn stdout_write(s: &str) -> Result<(), DownloadError> {
-    log(s);
-    Ok(())
 }
 
 /// Read up to 1 kB in memory.
